requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.7
django-anymail==11.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-nested-routers==0.94.1
drf-yasg==1.21.7
facebook-sdk==3.1.0
filelock==3.15.4
GDAL @ file:///C:/Users/Ericsson%20Raphael/Desktop/Natours_Django/GDAL-3.8.4-cp312-cp312-win_amd64.whl#sha256=841c9569814fb5167e5b5652cf244f20a91051d1d6dd1ee97e72a0a9c0631d8c
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
Js2Py==0.74
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pipwin==0.5.2
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.28.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyjsparser==2.7.1
PyJWT==2.9.0
pyparsing==3.1.4
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
setuptools==72.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.43.0